requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar
Django == 2.1.5
PyJWT == 1.7.1
djangorestframework == 3.9.2
